pyproject.toml
[project]
name = "echoxflow"
version = "0.2.1"
description = "Library and example tasks for reading and loading the EchoXFlow echocardiographic dataset"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
  "av>=12.0.0",
  "matplotlib>=3.9.0",
  "numpy>=2.0.0",
  "pillow>=10",
  "pydicom>=3.0.0",
  "PyYAML>=6.0.1",
  "tqdm>=4.67.3",
  "zarr>=3.1",
]

[project.optional-dependencies]
dev = [
  "black==25.1.0",
  "flake8==7.3.0",
  "flake8-pyproject==1.2.3",
  "isort==6.0.1",
  "mypy==1.14.1",
  "pre-commit==4.2.0",
  "pytest==8.3.5",
  "torch>=2.0.0",
  "types-PyYAML==6.0.12.20250402",
  "yamllint==1.37.1",
]
tasks = [
  "monai>=1.3.0",
  "torch>=2.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
echoxflow = ["config/*.yml"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
tag_format = "v{version}"
commit_parser = "conventional"
build_command = ""
upload_to_vcs_release = true
allow_zero_version = true

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 120
py_version = 313
skip_gitignore = true

[tool.flake8]
max-line-length = 120
select = ["E", "F", "W", "C90"]
extend-ignore = ["E203", "W503"]
exclude = [
  ".git",
  "__pycache__",
  ".venv",
  ".uv-cache",
  ".mypy_cache",
  ".pytest_cache",
  "build",
  "dist",
  "outputs",
  "tmp",
  ".tmp",
  "tasks/*/artifacts",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
exclude = [
  "^\\.venv/",
  "^\\.uv-cache/",
  "^\\.mypy_cache/",
  "^\\.pytest_cache/",
  "^outputs/",
  "^tmp/",
  "^\\.tmp/",
  "^tasks/.*/artifacts/",
]

[[tool.mypy.overrides]]
module = "av.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "matplotlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "monai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydicom.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tqdm.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
  "black==25.1.0",
  "flake8==7.3.0",
  "flake8-pyproject==1.2.3",
  "isort==6.0.1",
  "mypy==1.14.1",
  "pre-commit==4.2.0",
  "pytest==8.3.5",
  "torch>=2.0.0",
  "types-PyYAML==6.0.12.20250402",
  "yamllint==1.37.1",
]
tasks = [
  "monai>=1.3.0",
  "torch>=2.0.0",
]
